% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OData.R
\name{downloadResourceCsv}
\alias{downloadResourceCsv}
\title{Download a CSV file from an URL}
\usage{
downloadResourceCsv(url, separator)
}
\arguments{
\item{url}{URL pointing at a CSV file}

\item{separator}{Separator for the CSV file}
}
\value{
Data from a CSV file
}
\description{
Download a CSV file from an URL
}
\author{
BPM-Conseil
}

