% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OData.R
\name{entitySets}
\alias{entitySets}
\title{Get a list of EntitySets for a web service base on OData}
\usage{
entitySets(url)
}
\arguments{
\item{url}{URL to OData (needs to include $format=json as parameter)}
}
\value{
EnitySets of the OData target
}
\description{
Get a list of EntitySets for a web service base on OData
}
\examples{
entitySets("http://services.odata.org/TripPinRESTierService/?$format=json")
}
\author{
BPM-Conseil
}

