% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OData.R
\name{metadata}
\alias{metadata}
\title{Get the metadata for a specific web service based on OData Protocol}
\usage{
metadata(url)
}
\arguments{
\item{url}{URL to OData ($metada can be included as parameter to get more informations)}
}
\value{
The metadata of the OData target
}
\description{
Get the metadata for a specific web service based on OData Protocol
}
\examples{
metadata("http://services.odata.org/TripPinRESTierService/$metadata")
}
\author{
BPM-Conseil
}

