% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OData.R
\docType{package}
\name{OData}
\alias{OData}
\alias{OData-package}
\title{OData: A helper package to access data from a web service based on OData Protocol.}
\description{
The OData package provides methods to easily call an OData web service and get data from datasets
}

