% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OData.R
\name{retrieveData}
\alias{retrieveData}
\title{Retrieve data as list from a dataset}
\usage{
retrieveData(url)
}
\arguments{
\item{url}{URL to dataset (needs to include $format=json as parameter)}
}
\value{
Data from the OData web service
}
\description{
Retrieve data as list from a dataset
}
\examples{
retrieveData("http://services.odata.org/TripPinRESTierService/Airports/?$format=json")
}
\author{
BPM-Conseil
}

